Fix Batik PDF rendering - use float instead of int for dimensions
Batik's LengthKey rejects Integer values for KEY_WIDTH/KEY_HEIGHT hints.
Change PDF_WIDTH_POINTS and PDF_HEIGHT_POINTS back to float type.
Add comment explaining the requirement to prevent future regressions.

+++ src/main/java/villagecompute/calendar/services/PDFRenderingService.java
@@ -24,8 +24,9 @@ public class PDFRenderingService {
 
     // PDF page size: 35w x 23h inches (landscape)
     // PDF units are in points (1 inch = 72 points)
-    private static final int PDF_WIDTH_POINTS = 35 * 72; // 2520 points
-    private static final int PDF_HEIGHT_POINTS = 23 * 72; // 1656 points
+    // NOTE: These MUST be float - Batik's LengthKey rejects Integer values
+    private static final float PDF_WIDTH_POINTS = 35 * 72; // 2520 points
+    private static final float PDF_HEIGHT_POINTS = 23 * 72; // 1656 points
 
     /**
      * Render SVG content to PDF format. Uses Apache Batik's PDFTranscoder for SVG to PDF
